Move main cmd to ./cmd
So that it can be run as a library.
